requirements.txt
openai
faiss-cpu
sentence-transformers
requests
pandas
numpy
python-dotenv
streamlit
reportlab
matplotlib
seaborn
tiktoken
pydantic
pytest
difflib
